requirements.txt
aws-cdk.core>=1.73.0
aws-cdk.aws-lambda>=1.73.0
aws-cdk.aws-s3>=1.73.0
aws-cdk.aws-batch>=1.73.0
aws-cdk.aws-ecs>=1.73.0
aws-cdk.aws-iam>=1.73.0
aws-cdk.aws-events>=1.73.0
aws-cdk.aws-events-targets>=1.73.0
yapf
